Restore PrettyPatch.
PrettyPatch was removed as part of an earlier cleanup patch[1] which dropped the Websites directory. For whatever reason, the PrettyPatch functionality was buried under Websites/bugs.webkit.org/PrettyPatch. This patch moves that code to Tools/Scripts/webkitruby/PrettyPatch, and updates the various scripts that rely on the old directory.

[1]: http://src.chromium.org/viewvc/blink?view=revision&revision=147526

R=jochen@chromium.org

Review URL: https://codereview.chromium.org/13646005