Remove the ENABLE_CSP_NEXT compile-time flag.
This patch drops the ~20 occurances of '#IFDEF ENABLE(CSP_NEXT)' from the
codebase. It shouldn't effect any web-visible behavior, as the interesting
bits of CSP 1.1 are still hidden away behind the runtime flag, whose behavior
is untouched by this patch.

The only non-mechanical change in this patch is the addition of an annotation
to the script@nonce attribute, ensuring that it's only exposed if the runtime
flag is set (that was an oversight in the original code).

R=eseidel@chromium.org, abarth@chromium.org

Review URL: https://codereview.chromium.org/13643007