Remove the ENABLE_IFRAME_SEAMLESS compile-time flag.
This patch drops the ~10 instances of '#IFDEF ENABLE(IFRAME_SEAMLESS)' from the
codebase. It shouldn't effect any web-visible behavior, as the interesting bits
are all hidden safely behind a runtime flag, whose behavior is untouched by
this patch.

The only non-mechanical change in this patch is the addition of an annotation
to the iframe@seamless attribute, ensuring that it's only exposed if the
runtime flag is set (that was an oversight in the original code).

R=eseidel@chromium.org

Review URL: https://codereview.chromium.org/13646006