Accelerated SVG reference filters are broken
The problem is that we rely on FilterEffectRenderer to build the filter graph, and we call backing()->updateFilters() before we have a FilterEffectRenderer.

The simplest fix right now is to call backing()->updateFilters() again after we have a FilterEffectRenderer. Before, when we called updateFilters() in RenderLayerBacking::updateGraphicsLayerGeometry, the call happened late enough to have a FilterEffectRenderer for SVG reference filters, but too late to fall back to software filter painting.
The long-term fix is to extract the SVG graph building functionality out of FilterEffectRenderer, and make it happen in RenderLayer::computeFilterOperations.

See WK Bug: https://bugs.webkit.org/show_bug.cgi?id=114035

BUG=227025

Review URL: https://codereview.chromium.org/13638015