Guard <link> beforeload against recursion.
The beforeload event on a link element can recurse if it mutates its
firing link element. Now guard against this, only allowing the
innermost (last!) change to run. This prevents multiple client
registration and stops the crash in the bug report (which is
reproduced in the test).

BUG=174920

Review URL: https://codereview.chromium.org/13725004