DevTools: [CPUProfile] return profile header as a result of stop command.
Originally the profile header was transferred via an event.
It works fine in ProfilesPanel especially when profile was recorded via console.profile()/console.profileEnd()
But if another panel, as example Timeline wants to record a profile. It will never receive the profile header.
So I extended stop method with profileHeader as the return value.

BUG=none
TEST=none
R=yurys@chromium.org, pfeldman@chromium.org

Review URL: https://codereview.chromium.org/13637005