Remove shouldPropagateCompositingToEnclosingFrame(), we always do
The Mac WebKit1 WebKit port supported compositing iframes into independent
native windows. Blink doesn't support this mode so this code is dead.
Also added a nicer getter for determining if a RenderLayerCompositor is for
the main frame or not since this is fairly commonly called.

Review URL: https://codereview.chromium.org/13838005