Repaint the slider container after setting the thumb position.
This ensures that everything gets repainted correctly, especially the
track for a <video>'s volume slider. Nothing else would invalidate the
track -- the painting of the track background happens specially in
RenderMediaControlsChromium.cpp:paintMediaVolumeSlider.

We do overpaint in other cases now, but for this type of control that
should not be a problem.

BUG=190880
R=ojan@chromium.org

Review URL: https://codereview.chromium.org/13622003