Remove now-unused volumeSliderOffsetFromMuteButton
This was only called by RenderMediaVolumeSliderContainer::layout(),
which was not used by Chromium/Blink. It was used for the vertical
offset for the volume slider; in the Mac port, the slider only
shows up when hovering over the mute button. Blink has a different
UI for media controls that doesn't use this function, and the caller
was removed in crrev.com/13884002.

R=ojan@chromium.org
BUG=

Review URL: https://codereview.chromium.org/13818023