Move bindings out of WebCore
To improve the organization of our code, this CL move the bindings out of the
WebCore directory. The bindings are a layer on top of WebCore that lets
JavaScript interact with WebCore.

This CL just moves the files and the gypi listing the files. A future CL will
create a bindings.gyp and move the build logic into this new directory. The
directory structure we are headed towards is described in:

https://docs.google.com/document/d/1BCRkA5ERZtBMXkXB1aMFKKLl5LR_Wq5LpAHtlEf1j_c/edit?usp=sharing

Review URL: https://codereview.chromium.org/13985002