Touch dragging should not send fake mouse down event as webapps may misinterpret it.
Before this patch, we use to make a call to EventHandler::handleMousePressEvent() on receiving a long press. This was done to have EventHandler setup appropriate state for the handleDrag() call that came next that would start the actualy drag drop. However, this sends an extra un-necessary mouse down event to the webpage. Mouse down is already sent once when we recieve a touch down. Long press should not be sending another mouse down. To fix this, we remove the handleMousePressEvent() call and do the neccessary EventHandler state setup manually before calling handleDrag().

BUG=168165


Review URL: https://chromiumcodereview.appspot.com/13913015