Simplify the patch to malloc() on systems using the system allocator.
This is a no-op on Mac but simplifies things on Win / Linux:
- Don't decorate classes with "operator new" etc. for WTF_MAKE_FAST_ALLOCATED.
- Don't check for NULL in fastMalloc() because tcmalloc has already done that.

I was expecting this to be noticeably perf positive but it comes out a wash on
DOM/*
If anyone knows of any tests particularly sensitive to malloc performance then
I'm happy to run them.

The patch does reduce binary bloat by a tiny amount, presumably because the
fastMalloc() NULL checks were inlined previously.
Before: 66894040 bytes (DRT)
After:  66889944 bytes
For a whopping saving of 4k!

R=eseidel@chromium.org,abarth@chromiumlorg
BUG=229735
TEST=wktry


Review URL: https://chromiumcodereview.appspot.com/14006002