[Android] Use build/apk_test.gypi to build gtest APKs
apk_test.gypi templatizes the steps for building such an APK, and
ensures that they are built in a consistent way (and that that "way"
is contained in a small component and easily changed).

This was blocked on the inability to include a different path in a
chromium build vs in a webkit build.

BUG=158821


Review URL: https://chromiumcodereview.appspot.com/13601020