Remove slow, useless tests that fail on trybots
dispatch-message-string-data.html is a test for a 1.5 old use-after-free
bug that didn't even reproduce the use-after-free when it was first submitted
and takes 10 seconds to run.

dfg-inline-resolve.html is testing JSC-specific optimization behavior
and is unlikely to catch any bugs in V8. And it's slow.

R=esprehn


Review URL: https://chromiumcodereview.appspot.com/14100002