Replace WorldContextHandle with DOMWrapperWorld in Callbacks
WorldContextHandle is leaky in isolated worlds, while
RefPtr<DOMWrapperWorld> has the proper weak semantics
to allow the associated v8::Context to be freed.

A future patch will replace the uses of WorldContextHandle in
EventListeners. However, those are not leaky in the same way
as Callbacks because we manually clear all event listeners on
navigation.

R=abarth
BUG=230460

Review URL: https://codereview.chromium.org/14109007