Fix presubmit checks so that we lint the TestExpectations files more often.
We were only checking for lint errors when we modified the TestExpectations
files themselves; however, it is possible for changes to other files under
LayoutTests/ to also introduce errors, so we should run the check more often.
Fortunately, linting the files is pretty fast.

R=jchaffraix@chromium.org
BUG=

Review URL: https://codereview.chromium.org/14130016