Remove lots of unneeded subclasses in <details>
We can remove all the subclasses of elements used to create <details>
to greatly simplify the code. This reduces the chance of errors where
we're using special elements in our implementation and is one step
in getting rid of <webkitShadowContent> which isn't needed anymore
since we always have Shadow DOM turned on.

R=dglazkov@chromium.org,morrita@chromium.org

Review URL: https://codereview.chromium.org/13991005