HTMLIdentifier should take less space
This CL teaches HTMLIdentifier that the localNames for HTMLNames and friends
are thread-safe and can be used on background threads. This lets us shrink down
HTMLIdentifier to a single String.

In a future CL, we can remove the HTMLIdentifier class entirely and simply have
a free function that returns a PassRef<StringImpl>.

Review URL: https://codereview.chromium.org/13942007