Make RenderObject::isBR not inline to fix the build
There's many callers of isBR that would now need to include
RenderText to make this work. For example RenderObject::isBeforeContent
and friends call isBR and would be broken by this change unless the cpp
files added RenderText.h

For now we should just make it a regular method call instead.

I compiled locally and didn't see any of the failures that appear now, but
it looks like perhaps it's related to the components build, debug, or
something else that made it work. We can consider making it inline again
when we untangle the header deps.

Review URL: https://codereview.chromium.org/14312004