Remove hit test from dispatchDecidePolicyForNavigationAction
There's no need to do a hit test + layout in order to decide the policy for a
given navigation action. The originatingNode parameter is only used by a few
test codepaths that printf() information. This hit test forces a layout at an
awkward time which can be expensive and/or not do the expected thing.

BUG=232770

Review URL: https://codereview.chromium.org/14328011