Ensure IDBDatabase manages IDBTransaction lifetime
Previously, the IDBDatabase maintained a raw pointer to its
IDBTransaction objects. Now that the transaction back-end to front-end
communication is via the IDBDatabase, this could allow the
IDBTransaction wrapper to be GC'd before an abort or complete message
arrives, releasing the IDBTransaction itself, leading to a crash.

This patch changes it to a RefPtr; this avoids cycles because the
IDBTransaction notifies the IDBDatabase on completion which breaks the
cycle.

The unit test turned up a subtle but apparently harmless quirk caused
by arbitrary ordering of ActiveDOMObject::stop() calls; if if the
transaction was stopped first it would abort the request leading to an
enqueued event and dangling ref not cleaned up until the event queue
was destructed. If the request was stopped first this would be
skipped. Changed IDBTransaction to not abort requests if the context
is stopped for consistency.

BUG=226090

Review URL: https://chromiumcodereview.appspot.com/14236002