Make Frame's ScriptController an OwnPtr and remove the #include
Merges http://trac.webkit.org/changeset/148373 by AwesomeWeinig
(but the projects have already diverged enough that the actual diff
looks pretty different).

Frame is included in almost 1000 files, so keeping it small speeds up
compilation.

This exposes a null-pointer deref bug in XMLHttpRequest that
happened to work until now. To fix, also merge Sam's
http://trac.webkit.org/changeset/148380:
"Check that the frame is not null (as it can be in cases like
http/tests/xmlhttprequest/detaching-frame-2.html).
We used to be getting lucky, in that shouldBypassMainWorldContentSecurityPolicy(),
the function that is ultimately called, only operates on global state. Now that we need
to actually dereference the Frame to get the ScriptController, we see this crash."

BUG=none

Committed: https://src.chromium.org/viewvc/blink?view=rev&revision=148570

Review URL: https://codereview.chromium.org/14323004