Stop calling webkit-build-directory to figure out the build dir and config.
This code was used to make webkitpy work with wekbit-build-directory
and set-webkit-configuration, neither of which work in blink any more.
This lets us get rid of a lot of scaffolding as well.

This change fixes 'git cl upload' on windows, which is currently crashing
in a presubmit check trying to lint the test expectations (which ends up
calling webkit-build-directory and choking in some cases).

R=abarth@chromium.org, eseidel@chromium.org, dglazkov@chromium.org

Review URL: https://codereview.chromium.org/14100010