Remove frame flattening support as Chromium has no intention of using it
Chromium does not use the frame flattening code on any platform, though it used
to use it for frameset flattening on Android, but that is no longer the case.

As there is no intention of ever using it, it should be removed.

Committed on behalf of Kenneth Rohde Christiansen <kenneth.r.christiansen@intel.com>

BUG=231898

Review URL: https://codereview.chromium.org/14076011