Delete FrameLoader::m_policyDocumentLoader.
All policy callbacks are synchronously, so we can rely on the RefPtrs on the
stack to keep the DocumentLoader alive between creation and becoming the
provisional document loader.

We need to call DocumentLoader::setFrame directly if the DocumentLoader will
be used, since setPolicyDocumentLoader used to do that.

We also need to give FrameLoaderClient::decidePolicyForNavigationAction a
DocumentLoader now.

BUG=none
TEST=none

Review URL: https://codereview.chromium.org/14189008