Do not reuse ExclusionShapeInsideInfo in a recursive inline layout inside a region
ExclusionShapeInsideInfo cannot be used by inline descendants of elements that
take part in inline layout, such as floats or inline blocks. This behavior was
already implemented in the non-regions case, this patch just fixes it in the case
where an ExclusionShapeInsideInfo is picked up from a region.

R=jchaffraix@chromium.org
BUG=233327

Review URL: https://chromiumcodereview.appspot.com/14320007