Remove dead code about overlap testing from RenderLayerCompositor
In RenderLayerCompositong::computeCompositingRequirements, overlap testing was
always being used. The setting for using overlap testing was being initialized
to true, and never changed anywhere in the code. This patch removes that
boolean flag, and cleans up the if-statements and branching logic that were
used by that flag.

BUG=233466

Review URL: https://chromiumcodereview.appspot.com/14362020