Remove almost all of PageGroup
Almost all of the code in PageGroup is unused by Chromium
because we use a different facility for handling visited
link coloring and the rest of the code was to support
TestRunner features that we never implemented or used.

We no longer need a map of named PageGroups because Chromium
only uses private (formerly single) ones and a single named
one "default". Instead we can just create two kinds of groups,
a shared one that regular web pages are in, and private per
Page ones that are special cased like the web inspector. Making
this change allows a bunch of simplification.

R=abarth@chromium.org,

Review URL: https://codereview.chromium.org/14366020