Replace isXMLMimeType regex with a manual parser
We can replace the regex for matching +xml mimetypes with a manual
parser removing the need for thread safe storage of the regex which
was needed since you can create XMLHttpRequests from workers.

This is also neccesary to switch RegularExpression to be v8 powered
and not need to create a new v8::Context for every worker thread.

Review URL: https://codereview.chromium.org/14389002