Fixes for serializing SkPictures.
LazyDecodingPixelRef:
Do not return the original encoded data if the image has been scaled
and/or clipped, since the playback does not know how to reproduce
those operations.

ImageFrameGenerator:
Provide an accessor to find the full size of the image. Used by
LazyDecodingPixelRef to determine whether it is scaled.

BUG=168220

Review URL: https://chromiumcodereview.appspot.com/14099008