Fix the grid-{end|after} position resolution
Our resolution code was resolving both positions in the same way
but that doesn't work: in grid-rows: 2 / 3, the 2 lines, even if
they are different numbers, refers to the same grid position so
using the same formula to resolve them cannot work.

Because of the previous issue, we were off-by-one when resolving
grid-{end|after}. This only fixes positive explicit position as
the negative case needs more testing and scaffolding.

BUG=231695
TEST=fast/css-grid-layout/grid-item-spanning-resolution.html

Review URL: https://codereview.chromium.org/14246016