Add MeasureAs for window.event.
This adds a histogram to measure how often window.event is being used. The
intent of this is to determine whether we can remove this old IE specific
property.

This also makes window.event non enumerable so that it is not included when
people enumerate over the properties of the window object.

BUG=232371

Review URL: https://codereview.chromium.org/14034017