Simplify and improve BaselineOptimizer.
Post-fork we have a much simpler fallback order for test baselines.
Mainly, there's only one directory (LayoutTests) that is a meta
directory of subdirectories (chromium-mac and chromium-win). That is
to say, it's the only case where baselines earlier in the fallback order can
override the parent's baseline, i.e. if chromium-mac and chromium-win
have the same baseline, then that baseline overwrites the LayoutTests baseline.

As such, we can get a fully optimal baseline by remove redundant
baselines earlier in the fallback order and then checking if
chromium-mac and chromium-win have the same baseline.
This makes the code much simpler and always finds the optimal result.

Review URL: https://codereview.chromium.org/14044016