Switched to use HIDDEN_PAGE_DOM_TIMER_THROTTLING feature unconditionally. Removed per-page control over minimum DOM timer intervals; this is subsumed by the new feature. Part of this change reverts http://wkbug.com/54312 ; part switches over to the new feature from http://wkbug.com/98474 and simplifies it slightly.
Removed the ability to toggle throttling per-page; it is unconditionally enabled. The complexity of controlling it from the embedder is undesirable.

More code will be removed once this lands and is rolled in.

Did full layout test run with this change. New test passes. Verified manually that, on hidden pages, requestAnimationFrame callbacks still do not run at all, and that setTimeout callbacks are throttled.

BUG=227153
TEST=fast/dom/timer-throttling-hidden-page.html
R=esprehn@chromium.org,jamesr@chromium.org

Review URL: https://codereview.chromium.org/14126002